Move api calls to api.ts from daily.ts
