+++ README.md
@@ -1,22 +1,3 @@
 # ZACHROBERTSON.TECH
 
-This is my website built using [Next.js](https://www.nextjs.com/).
-
-You can find the old versions of the site at `branchName.zachrobertson.tech`
-
-Current Version: v1
-
-Old Versions:
-
-- [v0](https://github.com/zachrobertson/zachrobertson.tech/tree/v0)
-
-- [v1](https://github.com/zachrobertson/zachrobertson.tech/tree/v1)
-
-## Layout
-
-There is a standard `pageLayout` component that each page will import and use as its base element. There are also two templates for the blog posts, one is for the abbreviated post containers on the home page and the blog page, while the other is the actual full blog that will be shown when a users navigates to a blog post.
-
-### Fonts
-
-- Uses the `Oswald` font from the `@fontsource/oswald` npm package
-- All fonts sizes should be defined in `rem` units (the base root em size is 16px)
+This is my website built using [Next.js](https://www.nextjs.com/) and deployed on [Vercel](https://www.vercel.com)
