+++ src/lib/api.ts
@@ -4,7 +4,8 @@ import matter from "gray-matter";
 import { Parser } from "htmlparser2";
 import { BlogData } from "@/interfaces/blog";
 import markdownToHtml from "./markdownToHtml";
-import { StravaWeeklyStats } from "@/interfaces/strava";
+import { fetchAthleteActivities } from "@/lib/stravaApiRequests";
+import { StravaActivityRequestParams, StravaWeeklyStats } from "@/interfaces/strava";
 
 const BLOGS_DIRECTORY = path.join(process.cwd(), "_blogs");
 const EXCERPT_LENGTH = 500;
@@ -57,6 +58,55 @@ function _truncateHTML(html: string, length: number): string {
     return truncatedHtml;
 }
 
+const TWENTY_FOUR_WEEKS_IN_SECONDS = 48 * 7 * 24 * 60 * 60;
+
+function getWeek(date: Date): string {
+    date.setHours(0, 0, 0, 0);
+    // Thursday in current week decides the year.
+    date.setDate(date.getDate() + 3 - (date.getDay() + 6) % 7);
+    // January 4 is always in week 1.
+    var week1 = new Date(date.getFullYear(), 0, 4);
+    // Adjust to Thursday in week 1 and count number of weeks from date to week1.
+    const weekNumber = 1 + Math.round(((date.getTime() - week1.getTime()) / 86400000
+                            - 3 + (week1.getDay() + 6) % 7) / 7);
+    const year = date.getFullYear();
+    return `${year}-W${weekNumber}`;
+}
+
+
+function saveDataToFile(data: StravaWeeklyStats): void {
+    const filePath = path.join(process.cwd(), 'stravaData.json');
+    fs.writeFileSync(filePath, JSON.stringify(data));
+}
+
+export async function generateStravaDataFile() {
+    const before = Math.floor(new Date().getTime() / 1000);
+    const after = before - TWENTY_FOUR_WEEKS_IN_SECONDS;
+    const params: StravaActivityRequestParams = {
+        before,
+        after,
+        per_page: 200
+    };
+    const activities = await fetchAthleteActivities(params);
+
+    const weeklyStats: StravaWeeklyStats = {};
+    for (const activity of activities) {
+        if (activity.type == "Ride") {
+            const startDate = new Date(activity.start_date);
+            const weekIdx = getWeek(startDate);
+            const distance = activity.distance / 1609.34; // Distance in miles
+            const elevation = activity.total_elevation_gain / 0.3048; // Elevation in feet
+            if (!weeklyStats[weekIdx]) {
+                weeklyStats[weekIdx] = { miles: 0, elevation: 0 };
+            }
+            weeklyStats[weekIdx].miles += distance;
+            weeklyStats[weekIdx].elevation += elevation;
+        }
+    }
+
+    saveDataToFile(weeklyStats);
+}
+
 export async function getBlogById(id: string): Promise<BlogData> {
     const fileName = `${id}.md`
     const fullPath = path.join(BLOGS_DIRECTORY, fileName);
@@ -69,8 +119,7 @@ export async function getBlogById(id: string): Promise<BlogData> {
     if (data.headerImage == "stravaGraph") {
 
         if (!fs.existsSync("stravaData.json")) {
-            const baseUrl = process.env.NEXT_PUBLIC_BASE_URL || 'http://localhost:3000';
-            await fetch(`${baseUrl}/api/cron/daily`);
+            await generateStravaDataFile();
         }
     
         const rawData: StravaWeeklyStats = JSON.parse(fs.readFileSync("stravaData.json", 'utf-8'));
